// src/common/enums/application-status.enum.ts
// src/common/enums/application-status.enum.ts

export enum ApplicationStatus {
  APPLICATION_SUBMITTED = 'APPLICATION_SUBMITTED',
  BANK_VERIFICATION_PENDING = 'BANK_VERIFICATION_PENDING',
  PHONE_VERIFICATION_PENDING = 'PHONE_VERIFICATION_PENDING',
  DOCUMENT_REQUEST = 'DOCUMENT_REQUEST',
  SIGN_LOAN_AGREEMENT = 'SIGN_LOAN_AGREEMENT',
  VERIFICATION_DEPOSIT = 'VERIFICATION_DEPOSIT',
  FUNDED = 'FUNDED',
  DECLINED = 'DECLINED',
}